requirements.txt
flask
werkzeug
tensorflow
numpy
pillow